fix: tests fixed to use correct namedAccount, reordered params + tidy
